requirements.txt
pydantic==2.5.2
PyYAML==6.0.1
aiohttp==3.9.1
requests==2.31.0
psutil==5.9.6